cleaned up webpack config and dropped the .mjs extension for just js
