+++ src/js/app.js
@@ -1,9 +1,10 @@
-import { state } from "./state.mjs";
-import { pubsub } from "./pubsub.mjs";
-import { renderListTemplate } from "./renderlist.mjs";
-import { toggleDarkTheme, updateShaderUniforms } from "./eventhandler.mjs";
-import { note } from "./factories.mjs";
-import { drawCanvas } from "./canvas.mjs";
+import { state } from "./state.js";
+import { pubsub } from "./pubsub.js";
+import { renderListTemplate } from "./renderlist.js";
+import { toggleDarkTheme } from "./eventhandler.js";
+import { note } from "./factories.js";
+import { drawCanvas } from "./canvas.js";
+import { initPiano } from "./piano.js";
 
 state.notes.push(
   new note({ interval: 1, octave: 0, velocity: 1, color: [255, 0, 0] })
@@ -21,6 +22,7 @@ renderListTemplate(state);
 window.onload = (() => {
   drawCanvas();
 })
+initPiano();
 
 
 pubsub.subscribe("note added", renderListTemplate);
