--- src/js/constants.mjs
@@ -1,14 +1,0 @@
-const noteNames = [
-    "C",
-    "C#",
-    "D",
-    "Eb",
-    "E",
-    "F",
-    "F#",
-    "G",
-    "G#",
-    "A",
-    "Bb",
-    "B",
-];
