+++ src/js/state.js
@@ -1,0 +1,6 @@
+export let state = {
+  transpose: 0,
+  notes: [],
+  shaderUniforms: {},
+  theme: "light",
+};
